fixed arguments to snprintf to be always long long

+++ include/neso_particles/containers/cell_data.hpp
@@ -20,7 +20,7 @@ private:
   // std::format is C++20.......
   std::string format(INT value) {
     char buffer[128];
-    const int err = snprintf(buffer, 128, "%ld", value);
+    const int err = snprintf(buffer, 128, "%lld", static_cast<long long>(value));
     NESOASSERT(err >= 0 && err < 128, "Bad snprintf return code.");
     return std::string(buffer);
   }
